delete file feature created

+++ client/src/DragNDrop.js
@@ -10,6 +10,7 @@ class DnD {
   constructor() {
     this.trashZone = null;
     this.selectedId = null;
+    this.type = "";
     // this.drag = this.drag.bind(this);
     // this.dropInTrash = this.dropInTrash.bind(this);
     const methods = Object.getOwnPropertyNames(Object.getPrototypeOf(this));
@@ -24,8 +25,11 @@ class DnD {
     e.stopPropagation();
     const trashZone = document.getElementById("trash__zone");
     this.trashZone = trashZone;
+    this.type = e.currentTarget.dataset.type;
     this.trashZone.classList.add("delete__zone--over");
-    this.selectedId = e.currentTarget.dataset.folder_id;
+    this.selectedId =
+      e.currentTarget.dataset.folder_id || e.currentTarget.dataset.file_id;
+    console.log(e.currentTarget.dataset);
   }
 
   dragLeave() {
@@ -44,8 +48,6 @@ class DnD {
   dragEnd(e) {
     e.preventDefault();
     e.stopPropagation();
-    console.log("ended");
-    // console.log(this.trashZone);
     this.trashZone.classList.remove("delete__zone--over");
   }
 
@@ -53,21 +55,31 @@ class DnD {
   trashZoneDragOver(e) {
     e.preventDefault();
     this.trashZone.classList.add("delete__zone--over--dashed");
-    console.log("dragging");
   }
 
-  dropInTrash(e, pathKeys) {
+  dropInTrash(e, args) {
+    let { folderPaths, filePaths } = args;
     e.stopPropagation();
-    console.log("deleted");
     this.trashZone.classList.remove("delete__zone--over--dashed");
-    const path = pathKeys[this.selectedId].path;
+    console.log(filePaths);
+    const path =
+      this.type === "folder"
+        ? folderPaths[this.selectedId].path
+        : filePaths[this.selectedId].file_dir;
     this.deleteDirectoryApi(path);
     e.preventDefault();
   }
 
   async deleteDirectoryApi(path) {
     try {
-      await http.delete("/directories/delete", { data: { directory: path } });
+      console.log(path);
+      if (this.type === "folder")
+        await http.delete("/directories/delete", { data: { directory: path } });
+      if (this.type === "file")
+        await http.delete("/files/delete", { data: { file_dir: path } });
+
+      console.log(this.selectedId);
+      console.log(document.getElementById(this.selectedId));
       unmountComponent(this.selectedId);
       this.trashZone.classList.remove("delete__zone--over");
     } catch (error) {
